Code split edit view dialog
