+++ src/panels/lovelace/editor/show-edit-card-dialog.ts
